+++ plugin/client.go
@@ -1,6 +1,8 @@
 package plugin
 
 import (
+	"fmt"
+
 	"github.com/torlenor/abylebotter/model"
 )
 
@@ -9,13 +11,25 @@ type AbyleBotterPlugin struct {
 	// API exposes the plugin API of the bot.
 	API API
 
+	NeededFeatures []string
+
+	Type string
+
 	BotID    string
 	PluginID string
 }
 
 // SetAPI gives the API interface to the plugin.
-func (p *AbyleBotterPlugin) SetAPI(api API) {
+func (p *AbyleBotterPlugin) SetAPI(api API) error {
+	for _, f := range p.NeededFeatures {
+		if !api.HasFeature(f) {
+			return fmt.Errorf("Bot does not provided needed feature %s", f)
+		}
+	}
+
 	p.API = api
+
+	return nil
 }
 
 // SetBotPluginID sets the plugin ID to the given value
@@ -25,6 +39,9 @@ func (p *AbyleBotterPlugin) SetBotPluginID(botID string, pluginID string) {
 }
 
 // Default hook implementations (see hooks.go)
+
+// PluginType returns the plugin type
+func (p *AbyleBotterPlugin) PluginType() string { return p.Type }
 
 // OnRun in its default implementation
 func (p *AbyleBotterPlugin) OnRun() {}