+++ plugin/voteplugin/commandhandler.go
@@ -8,7 +8,7 @@ import (
 )
 
 const (
-	helpText        = "Type '!vote What is the best color? [Red, Green, Blue]' to start a new giveaway.\nYou can omit the custom options in the [...] to initiate a simple Yes/No vote."
+	helpText        = "Type '!vote What is the best color? [Red, Green, Blue]' to start a new vote.\nYou can omit the custom options in the [...] to initiate a simple Yes/No vote."
 	voteEndHelpText = "To end a vote type !voteend description text of the vote."
 )